LPS-54471 match kernel facet package
